Partial code and docs
Merging as the work is likely to be dropped for a while due to other
priorities but I want to keep it discoverable.
