+++ Pipfile
@@ -1,0 +1,12 @@
+[[source]]
+url = "https://pypi.python.org/simple"
+verify_ssl = true
+[packages]
+mongo-connector = "*"
+psycopg2 = "*"
+
+
+
+
+
+
