+++ mongo_connector/doc_managers/old/mapping_schema.py
