+++ mongo_connector/doc_managers/old/mappings.py
