+++ mongo_connector/doc_managers/old/utils.py
