+++ old/tests/test_postgresql_manager.py
